# docs/swagger.yaml
definitions:
...
  model.Product:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      title:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
  model.Role:
    properties:
      created_at:
        description: 创建时间
        type: string
      id:
        type: integer
      menu_id:
        description: 菜单ID    用逗号分隔例如 1，2，3，4
        type: string
      permissions_id:
        description: 权限ID    用逗号分隔例如 1，2，3，4
        type: string
      role_name:
        description: 角色名称
        type: string
      role_tag:
        description: 角色标签
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  response.Code:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
info:
  contact:
    email: xiangfudada@163.com
    name: Xfdada
  title: ICQQG接囗文档
  version: "1.0"
paths:
...
  /api/web/product:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      - description: 标题
        in: formData
        name: title
        required: true
        type: string
      - description: 型号
        in: formData
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 添加产品信息
      tags:
      - 后台产品模块
  /api/web/product/{id}:
    delete:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 删除产品信息
      tags:
      - 后台产品模块
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":model.Product}'
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 获取产品信息
      tags:
      - 后台产品模块
  /api/web/product/{id}/edit:
    put:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      - description: 标题
        in: formData
        name: title
        required: true
        type: string
      - description: 型号
        in: formData
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 修改产品信息
      tags:
      - 后台产品模块
  /api/web/productList:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]model.Product}'
          schema:
            items:
              $ref: '#/definitions/model.Product'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 产品列表
      tags:
      - 后台产品模块
  /api/web/role:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 角色英文名
        in: formData
        name: role_name
        required: true
        type: integer
      - description: 角色中文名
        in: formData
        name: role_tag
        required: true
        type: string
      - description: 菜单ID字符串1，2，3
        in: formData
        name: menu_id
        required: true
        type: string
      - description: 权限ID字符串 1，2，3
        in: formData
        name: permissions_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 添加角色信息
      tags:
      - 后端角色模块
  /api/web/role/{id}:
    delete:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 删除角色信息
      tags:
      - 后端角色模块
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":model.Role}'
          schema:
            $ref: '#/definitions/model.Role'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 获取角色信息
      tags:
      - 后端角色模块
    put:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 角色英文名
        in: formData
        name: role_name
        required: true
        type: integer
      - description: 角色中文名
        in: formData
        name: role_tag
        required: true
        type: string
      - description: 菜单ID字符串1，2，3
        in: formData
        name: menu_id
        required: true
        type: string
      - description: 权限ID字符串 1，2，3
        in: formData
        name: permissions_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 修改菜单信息
      tags:
      - 后端菜单模块
  /api/web/roleList:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]model.Menu}'
          schema:
            items:
              $ref: '#/definitions/model.Role'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 角色列表
      tags:
      - 后端角色模块
...
swagger: "2.0"